Refactor much more tool testing code into galaxy-lib.
