+++ lib/galaxy/tools/verify/__init__.py
@@ -286,58 +286,3 @@ def files_contains(file1, file2, attributes=None):
             line_diff_count += 1
         if line_diff_count > lines_diff:
             raise AssertionError("Failed to find '%s' in history data. (lines_diff=%i):\n" % (contains, lines_diff))
-
-
-# Galaxy specific methods - rest of this can be used with arbitrary files and such.
-
-def verify_hid(filename, hda_id, attributes, test_data_path_builder, hid="", dataset_fetcher=None, keep_outputs_dir=False):
-    assert dataset_fetcher is not None
-
-    def verify_extra_files(extra_files):
-        _verify_extra_files_content(extra_files, hda_id, dataset_fetcher=dataset_fetcher, test_data_path_builder=test_data_path_builder, keep_outputs_dir=keep_outputs_dir)
-
-    data = dataset_fetcher(hda_id)
-    item_label = "History item %s" % hid
-    verify(
-        item_label,
-        data,
-        attributes=attributes,
-        filename=filename,
-        get_filename=test_data_path_builder,
-        keep_outputs_dir=keep_outputs_dir,
-        verify_extra_files=verify_extra_files,
-    )
-
-
-def _verify_composite_datatype_file_content(file_name, hda_id, base_name=None, attributes=None, dataset_fetcher=None, test_data_path_builder=None, keep_outputs_dir=False):
-    assert dataset_fetcher is not None
-
-    data = dataset_fetcher(hda_id, base_name)
-    item_label = "History item %s" % hda_id
-    try:
-        verify(
-            item_label,
-            data,
-            attributes=attributes,
-            filename=file_name,
-            get_filename=test_data_path_builder,
-            keep_outputs_dir=keep_outputs_dir,
-        )
-    except AssertionError as err:
-        errmsg = 'Composite file (%s) of %s different than expected, difference:\n' % (base_name, item_label)
-        errmsg += str(err)
-        raise AssertionError(errmsg)
-
-
-def _verify_extra_files_content(extra_files, hda_id, dataset_fetcher, test_data_path_builder, keep_outputs_dir):
-    files_list = []
-    for extra_type, extra_value, extra_name, extra_attributes in extra_files:
-        if extra_type == 'file':
-            files_list.append((extra_name, extra_value, extra_attributes))
-        elif extra_type == 'directory':
-            for filename in os.listdir(test_data_path_builder(extra_value)):
-                files_list.append((filename, os.path.join(extra_value, filename), extra_attributes))
-        else:
-            raise ValueError('unknown extra_files type: %s' % extra_type)
-    for filename, filepath, attributes in files_list:
-        _verify_composite_datatype_file_content(filepath, hda_id, base_name=filename, attributes=attributes, dataset_fetcher=dataset_fetcher, test_data_path_builder=test_data_path_builder, keep_outputs_dir=keep_outputs_dir)
